+++ lib/service/translate/index.dart
@@ -2,12 +2,14 @@ import 'dart:core';
 
 import 'package:anx_reader/config/shared_preference_provider.dart';
 import 'package:anx_reader/enums/lang_list.dart';
+import 'package:anx_reader/service/translate/deepl.dart';
 import 'package:anx_reader/service/translate/google.dart';
 import 'package:anx_reader/service/translate/microsoft.dart';
 
 enum TranslateService {
   google('Google'),
-  microsoft('Microsoft');
+  microsoft('Microsoft'),
+  deepl('DeepL');
 
   const TranslateService(this.label);
 
@@ -77,6 +79,8 @@ class TranslateFactory {
         return GoogleTranslateProvider();
       case TranslateService.microsoft:
         return MicrosoftTranslateProvider();
+      case TranslateService.deepl:
+        return DeepLTranslateProvider();
     }
   }
 }